+++ Source/Core/Core/PowerPC/Jit64IL/IR_X86.cpp
@@ -1665,9 +1665,9 @@ static void DoWriteCode(IRBuilder* ibuild, JitIL* Jit, u32 exitAddress) {
 			X64Reg reg = regFindFreeReg(RI);
 			u64 val = ibuild->GetImmValue64(I);
 			if ((u32)val == val)
-				Jit->MOV(32, R(reg), Imm32(val));
+				Jit->MOV(32, R(reg), Imm32((u32)val));
 			else if ((s32)val == (s64)val)
-				Jit->MOV(64, R(reg), Imm32(val));
+				Jit->MOV(64, R(reg), Imm32((s32)val));
 			else
 				Jit->MOV(64, R(reg), Imm64(val));
 			RI.regs[reg] = I;
